# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GoleanAuth API
  version: 1.0.0
  summary: Production-grade authentication service.
  description: |
    GoleanAuth is a production-ready authentication service supporting
    email/password registration and login, Sign in with Google and Apple,
    refresh token rotation, session management, an OAuth 2.0 / OIDC provider
    surface (authorization code with PKCE, client credentials, introspection,
    revocation, userinfo, discovery, JWKS), admin client registration, and
    audit logging.

  contact:
    name: Miracle Chukwu
    email: support@goleanauth.dev

  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local Development

  - url: https://api.goleanauth.com
    description: Production
...
paths:
...
  /v1/auth/refresh:
    post:
      tags: [Sessions]
      summary: Refresh the access token
      description: |
        Rotates the refresh token presented in the `refresh_token` HttpOnly
        cookie and returns a fresh access token.
      operationId: refreshToken
      security:
        - refreshCookie: []
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
        "401":
          description: Missing or invalid refresh token
...
  /login:
    get:
      tags: [OAuth]
      summary: Interactive login form
      description: HTML page used by the authorize flow. Rendered when the browser has no valid session.
      operationId: loginForm
      parameters:
        - name: next
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: HTML login form
    post:
      tags: [OAuth]
      summary: Submit the interactive login form
      operationId: loginSubmit
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                password:
                  type: string
                next:
                  type: string
              required: [identifier, password]
      responses:
        "302":
          description: Redirect back to the authorize flow (refreshed cookie set)
        "200":
          description: Re-rendered form with an error
  /v1/oauth/authorize:
    get:
      tags: [OAuth]
      summary: Authorize a client on behalf of the user
      description: |
        Interactive OAuth 2.0 authorization endpoint. Validates the request,
        then shows the consent page (or redirects to /login when the browser
        has no session). Supports PKCE (RFC 7636) via code_challenge.
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: false
          schema:
            type: string
            enum: [S256]
      responses:
        "200":
          description: Consent page HTML
        "307":
          description: Redirect to /login when unauthenticated
  /v1/oauth/approve:
    post:
      tags: [OAuth]
      summary: Submit the consent decision
      description: |
        Approve or deny an authorization request. On approval a single-use
        authorization code is issued and the browser is redirected to the
        client's redirect_uri with `code` and `state`.
      operationId: consentApprove
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                redirect_uri:
                  type: string
                scope:
                  type: string
                state:
                  type: string
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                decision:
                  type: string
                  enum: [approve, deny]
              required: [client_id, redirect_uri, decision]
      responses:
        "302":
          description: Redirect to redirect_uri (code or access_denied)
        "200":
          description: Error page for invalid requests
  /v1/oauth/token:
    post:
      tags: [OAuth]
      summary: Exchange a grant for tokens
      description: |
        Issues access and refresh tokens for the client_credentials,
        authorization_code (with optional PKCE verification), and
        refresh_token grants. Clients authenticate with HTTP Basic or form
        client_id/client_secret.
      operationId: token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum: [client_credentials, authorization_code, refresh_token]
                client_id:
                  type: string
                client_secret:
                  type: string
                code:
                  type: string
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
                scope:
                  type: string
              required: [grant_type]
      responses:
        "200":
          description: Token set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: invalid_request, invalid_grant, invalid_scope, or unsupported_grant_type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenError"
        "401":
          description: invalid_client
  /v1/oauth/revoke:
    post:
      tags: [OAuth]
      summary: Revoke a refresh token
      description: RFC 7009 revocation. Always returns 200 once the client is authenticated.
      operationId: revoke
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                client_secret:
                  type: string
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Token revoked
  /v1/oauth/introspect:
    post:
      tags: [OAuth]
      summary: Introspect a token
      description: RFC 7662 introspection for access and refresh tokens.
      operationId: introspect
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                client_secret:
                  type: string
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Token state and claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectionResponse"
  /v1/userinfo:
    get:
      tags: [OAuth]
      summary: Standard user claims
      description: OpenID Connect userinfo for the authenticated principal.
      operationId: userinfo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Invalid token
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key

    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
        refresh_token:
          type: string
        scope:
          type: string
    TokenError:
      type: object
      properties:
        error:
          type: string
          enum: [invalid_request, invalid_client, invalid_grant, invalid_scope, unsupported_grant_type]
        error_description:
          type: string
    IntrospectionResponse:
      type: object
      properties:
        active:
          type: boolean
        sub:
          type: string
        session_id:
          type: string
        client_id:
          type: string
        scope:
          type: string
        iss:
          type: string
        aud:
          type: array
          items:
            type: string
        exp:
          type: integer
        iat:
          type: integer
    UserInfo:
      type: object
      properties:
        sub:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        given_name:
          type: string
        family_name:
          type: string
        name:
          type: string
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
...